Fix what a full review of the February cleanup work found
The second slice of the branch review: the fifty-three February commits that
fixed undefined variables, missing methods, mismatched names, duplicate keys
and the like across the components, and reworked the config repository.
Read from three angles - line by line, removed behaviour, and callers - with
every candidate verified against the code and, where it could be, live.

Config

1. Per-client configuration was no longer loaded. The repository used to
   take the client and pass it to the loader, which merges config/<client>/
   over the base; the rewrite passed nothing, so an api/ or administrator/
   override directory was silently ignored. The client is passed again.

2. The Joomla-era configuration.php was looked for under app/ rather than
   the site root, so a hub not yet migrated to config/ booted with an empty
   configuration. The loader's root path is PATH_ROOT again.

3. Saving Members > Registration with caching on threw: the repository now
   refuses a bare key that is not a config group, and the cache handler name
   is not one. Dot-notation, as com_cache was already changed to use. The
   set() docblock described a fallback the code does not have; it now says
   why the throw is the right answer.

Components and plugins

4. The storefront image form rendered a layout that has never existed - not
   on this branch, not before it - so the noscript fallback and every
   non-ajax upload result died. The view exists now and displayTask() reads
   the object from the query string the edit pages put there.

5. Re-subscribing a newsletter address called load($id) on a Relational
   model, which has no such fetch; the row was never loaded and save()
   inserted a new one. It uses oneOrFail() like the tasks around it.

6. The project link preview asked voku's find($sel, 0) whether a meta
   description exists; it returns a blank node, never null, so the body-text
   fallback was dead. findOneOrFalse() and getAttribute().

7. Deleting a project feed comment resolved the wrong id: the link says cid,
   the handler read activity. It accepts both.

8. The group Collections "Following" tab showed followers; there was no
   following() at all, on this branch or before. Written from followers()
   and the members plugin's equivalent, for the following template that was
   already there.

9. Re-saving a tool's dev group made every developer one of its managers -
   and a manager can add members, that is developers, which is the tool
   admin's decision. The managers the group has are left alone.

10. Sharing a tool session copied connection details from the owner's
    viewperm row without checking one exists. Guarded.

11. The forum API looked up a thread root by object and scope with no parent
    filter and no ordering, so a reply could stand in for the root and the
    tree walk started in the wrong place. The root is the post with no
    parent, in both lookups.

12. The FTP adapter's deleteDirectory() listed bare names and deleted them
    relative to the session's working directory, and accepted a $preserve
    flag it ignored. Full paths; preserve honoured as Local does.

13. The jobs API carried an authorization helper nothing called, on a class
    with no can() - dead code that would fatal the moment it was wired in.
    Removed.

14. The promo opt-out built Awards from User::get('id'); the constructor
    treats anything that is not an integer as a profile object. Cast.

15. The InstallTemplateEntry macro's signature gained two parameters in the
    middle; a migration written to the old five-argument form would have put
    its styles in the enabled slot. The old form is recognised and mapped.

Found live while checking: an empty search hit a preg_replace() deprecation
on every request (com_search terms, from January, outside this slice). Cast.

Checked and found sound: the Repository throw on an unknown bare key (tested
design; the shipped app.php carries every key the plugins set); the KB
comments feed link (Document::setLink() runs first); the support ticket
severity default (nothing filters on it); the members media admin check
(component-scoped, as the rest of com_members); decoupling plugins and
modules from Obj (deliberate; no in-tree or hub-local caller depends on it);
the plgX-to-namespace rewrites; the language-pack restructure (no key lost);
the API controller renames; the vendored markdown and DOM swaps.

Verified: lint and PSR-12 clean on every touched file; 113 config tests, 41
install/menu tests and the 2,905-test configured suite unchanged; site,
administrator and API clients serving on all three hubs; the 78-url probe
unchanged against baseline; routes clean on all three hubs.

+++ core/components/com_members/site/controllers/profiles.php
@@ -69,7 +69,7 @@ class Profiles extends SiteController
             return;
         }
 
-        $ia = new \Components\Members\Models\Incremental\Awards(User::get('id'));
+        $ia = new \Components\Members\Models\Incremental\Awards((int) User::get('id'));
         $ia->optOut();
 
         App::redirect(
